Checkout be

See merge request dev-la/milc-api!28

+++ app/DataTransformer/Product/ProductUpdateDataTransformer.php
@@ -15,7 +15,6 @@ class ProductUpdateDataTransformer
     {
         DB::beginTransaction();
 
-        $availableFormatsRequest = $genresRequest = [];
         if (isset($arrayRequest['available_formats'])) {
             $availableFormatsRequest = $arrayRequest['available_formats'];
             unset($arrayRequest['available_formats']);
@@ -26,70 +25,59 @@ class ProductUpdateDataTransformer
             unset($arrayRequest['genres']);
         }
 
-        $dubFilesRequest = [];
         if (isset($arrayRequest['dub_files'])) {
             $dubFilesRequest = $arrayRequest['dub_files'];
             unset($arrayRequest['dub_files']);
         }
 
-        $subtitlesRequest = [];
         if (isset($arrayRequest['subtitles'])) {
             $subtitlesRequest = $arrayRequest['subtitles'];
             unset($arrayRequest['subtitles']);
         }
 
-        $promotionalVideosRequest = [];
         if (isset($arrayRequest['promotional_videos'])) {
             $promotionalVideosRequest = $arrayRequest['promotional_videos'];
             unset($arrayRequest['promotional_videos']);
         }
-        
-        
-        $rightsBundlesRequest = [];
-        
+
         if (isset($arrayRequest['rights_bundles'])) {
             $rightsBundlesRequest = $arrayRequest['rights_bundles'];
         }
-        
-        $productionInfoRequest = [];
 
         if (isset($arrayRequest['production_info'])) {
             $productionInfoRequest = $arrayRequest['production_info'];
         }
 
-        if (isset($productionInfoRequest['directors'])) {
-            $directorsRequest = $productionInfoRequest['directors'];
-            unset($productionInfoRequest['directors']);
-        }
-        if (isset($productionInfoRequest['producers'])) {
-            $producersRequest = $productionInfoRequest['producers'];
-            unset($productionInfoRequest['producers']);
-        }
-        if (isset($productionInfoRequest['writers'])) {
-            $writersRequest = $productionInfoRequest['writers'];
-            unset($productionInfoRequest['writers']);
-        }
-        if (isset($productionInfoRequest['cast'])) {
-            $castRequest = $productionInfoRequest['cast'];
-            unset($productionInfoRequest['cast']);
+        if (isset($productionInfoRequest)) {
+            if (isset($productionInfoRequest['directors'])) {
+                $directorsRequest = $productionInfoRequest['directors'];
+                unset($productionInfoRequest['directors']);
+            }
+            if (isset($productionInfoRequest['producers'])) {
+                $producersRequest = $productionInfoRequest['producers'];
+                unset($productionInfoRequest['producers']);
+            }
+            if (isset($productionInfoRequest['writers'])) {
+                $writersRequest = $productionInfoRequest['writers'];
+                unset($productionInfoRequest['writers']);
+            }
+            if (isset($productionInfoRequest['cast'])) {
+                $castRequest = $productionInfoRequest['cast'];
+                unset($productionInfoRequest['cast']);
+            }
         }
 
         unset($arrayRequest['production_info']);
-
-        $marketingAssetsRequest = [];
 
         if (isset($arrayRequest['marketing_assets'])) {
             $marketingAssetsRequest = $arrayRequest['marketing_assets'];
             unset($arrayRequest['marketing_assets']);
         }
 
-        $productionImagesRequest = [];
-        if (isset($marketingAssetsRequest['production_images'])) {
+        if (isset($marketingAssetsRequest) && isset($marketingAssetsRequest['production_images'])) {
             $productionImagesRequest = $marketingAssetsRequest['production_images'];
             unset($marketingAssetsRequest['production_images']);
         }
-
-        $rightsInformationRequest = [];
 
         if (isset($arrayRequest['rights_information'])) {
             $rightsInformationRequest = $arrayRequest['rights_information'];
@@ -99,54 +87,59 @@ class ProductUpdateDataTransformer
         $productRequest = $arrayRequest;
 
         $productionInfo = $product->production_info()->first();
-        $productionInfo->update($productionInfoRequest);
+
+        if (isset($productionInfoRequest)) {
+            $productionInfo->update($productionInfoRequest);
+        }
 
         $productRequest['production_info_id'] = $productionInfo->id;
-        
+
         if (isset($rightsBundlesRequest)) {
             $product->rights_bundles()->detach();
             foreach ($rightsBundlesRequest as $bundleRightRequest) {
-                
-                $rightsBundlesRightsInfoRequest = [];
+
                 if (isset($bundleRightRequest['rights_information'])) {
                     $rightsBundlesRightsInfoRequest = $bundleRightRequest['rights_information'];
                     unset($bundleRightRequest['rights_information']);
                 }
-                
+
                 $priceRequest = [];
                 if (isset($bundleRightRequest['price'])) {
                     $priceRequest = $bundleRightRequest['price'];
                     unset($bundleRightRequest['price']);
                 }
-                
+
                 if (isset($priceRequest['id'])) {
                     $price = Money::findOrFail($priceRequest['id']);
                     $price->update($priceRequest);
                 } else {
                     $price = Money::create($priceRequest);
                 }
-                
+
                 $price->save();
-                
+
                 $bundleRightRequest['price_id'] = $price->id;
-                
+                $bundleRightRequest['product_id'] = $product->id;
+
                 if (isset($bundleRightRequest['id'])) {
                     $bundleRight = RightsBundle::findOrFail($bundleRightRequest['id']);
                     $bundleRight->update($bundleRightRequest);
                 } else {
                     $bundleRight = RightsBundle::create($bundleRightRequest);
                 }
-                
+
                 $bundleRight->save();
-                
-                $bundleRight->bundle_rights_information()->detach();
-                
-                foreach($rightsBundlesRightsInfoRequest as $rightsBundlesRightInfoRequest){
-                    $bundleRight->bundle_rights_information()->attach($rightsBundlesRightInfoRequest);
+
+                if (isset($rightsBundlesRightsInfoRequest)) {
+                    $bundleRight->bundle_rights_information()->detach();
+
+                    foreach ($rightsBundlesRightsInfoRequest as $rightsBundlesRightInfoRequest) {
+                        $bundleRight->bundle_rights_information()->attach($rightsBundlesRightInfoRequest);
+                    }
                 }
-                
+
                 $bundleRight->save();
-                
+
                 $product->rights_bundles()->attach($bundleRight->id);
             }
         }
@@ -205,81 +198,108 @@ class ProductUpdateDataTransformer
 
         $marketingAssets = $product->marketing_assets;
 
-        $marketingAssets->production_images()->detach();
-        foreach ($productionImagesRequest as $productionImageId) {
-            $marketingAssets->production_images()->attach($productionImageId);
+        if (isset($productionImagesRequest)) {
+            $marketingAssets->production_images()->detach();
+
+            foreach ($productionImagesRequest as $productionImageId) {
+                $marketingAssets->production_images()->attach($productionImageId);
+            }
         }
 
-        $marketingAssets->update($marketingAssetsRequest);
+        if (isset($marketingAssetsRequest)) {
+            $marketingAssets->update($marketingAssetsRequest);
+        }
 
         $productRequest['marketing_assets_id'] = $marketingAssets->id;
 
         $rightsInformationArray = [];
-        foreach ($rightsInformationRequest as $rightsInformationRequestItem) {
+        if (isset($rightsInformationRequest)) {
+            foreach ($rightsInformationRequest as $rightsInformationRequestItem) {
 
-            if (isset($rightsInformationRequestItem['available_rights'])) {
-                $availableRightsRequest = $rightsInformationRequestItem['available_rights'];
-                unset($rightsInformationRequestItem['available_rights']);
-            }
+                if (isset($rightsInformationRequestItem['available_rights'])) {
+                    $availableRightsRequest = $rightsInformationRequestItem['available_rights'];
+                    unset($rightsInformationRequestItem['available_rights']);
+                }
 
-            $rightsInformation = null;
-            if (isset($rightsInformationRequestItem['id'])) {
-                $productRightsInformations = $product->rights_information()->get();
-                foreach ($productRightsInformations as $productRightsInformation) {
-                    if ($productRightsInformation->id === $rightsInformationRequestItem['id']) {
-                        $rightsInformation = $productRightsInformation;
+                $rightsInformation = null;
+                if (isset($rightsInformationRequestItem['id'])) {
+                    $productRightsInformations = $product->rights_information()->get();
+                    foreach ($productRightsInformations as $productRightsInformation) {
+                        if ($productRightsInformation->id === $rightsInformationRequestItem['id']) {
+                            $rightsInformation = $productRightsInformation;
+                        }
                     }
                 }
-            }
 
-            if (isset($rightsInformation) and $rightsInformation !== null) {
-                $rightsInformation->update($rightsInformationRequestItem);
-            } else {
-                $rightsInformation = RightsInformation::create($rightsInformationRequestItem);
-            }
-
-            if (isset($availableRightsRequest)) {
-                $rightsInformation->available_rights()->detach();
-                foreach ($availableRightsRequest as $availableRightId) {
-                    $rightsInformation->available_rights()->attach($availableRightId);
+                if (isset($rightsInformation) and $rightsInformation !== null) {
+                    $rightsInformation->update($rightsInformationRequestItem);
+                } else {
+                    $rightsInformation = RightsInformation::create($rightsInformationRequestItem);
                 }
+                
+                $rightsInformation->product_id = $product->id;
+
+                if (isset($availableRightsRequest) && count($availableRightsRequest) > 0) {
+                    $rightsInformation->available_rights()->detach();
+                    foreach ($availableRightsRequest as $availableRightId) {
+                        $rightsInformation->available_rights()->attach($availableRightId);
+                    }
+                }
+
+                $rightsInformationArray[] = $rightsInformation;
             }
-
-            $rightsInformationArray[] = $rightsInformation;
         }
-
+        
         $product->update($productRequest);
 
-        $product->available_formats()->detach();
-        foreach ($availableFormatsRequest as $availableFormatId) {
-            $product->available_formats()->attach($availableFormatId);
+        if (isset($availableFormatsRequest)) {
+            $product->available_formats()->detach();
+
+            foreach ($availableFormatsRequest as $availableFormatId) {
+                $product->available_formats()->attach($availableFormatId);
+            }
         }
 
-        $product->genres()->detach();
-        foreach ($genresRequest as $genreId) {
-            $product->genres()->attach($genreId);
+        if (isset($genresRequest)) {
+            $product->genres()->detach();
+
+            foreach ($genresRequest as $genreId) {
+                $product->genres()->attach($genreId);
+            }
         }
 
-        $product->dub_files()->detach();
-        foreach ($dubFilesRequest as $dubFileId) {
-            $product->dub_files()->attach($dubFileId);
+        if (isset($dubFilesRequest)) {
+            $product->dub_files()->detach();
+
+            foreach ($dubFilesRequest as $dubFileId) {
+                $product->dub_files()->attach($dubFileId);
+            }
         }
 
-        $product->subtitles()->detach();
-        foreach ($subtitlesRequest as $subtitlesId) {
-            $product->subtitles()->attach($subtitlesId);
+        if (isset($subtitlesRequest)) {
+            $product->subtitles()->detach();
+
+            foreach ($subtitlesRequest as $subtitlesId) {
+                $product->subtitles()->attach($subtitlesId);
+            }
         }
 
-        $product->promotional_videos()->detach();
-        foreach ($promotionalVideosRequest as $promotionalVideosId) {
-            $product->promotional_videos()->attach($promotionalVideosId);
+        if (isset($promotionalVideosRequest)) {
+            $product->promotional_videos()->detach();
+
+            foreach ($promotionalVideosRequest as $promotionalVideosId) {
+                $product->promotional_videos()->attach($promotionalVideosId);
+            }
         }
 
-        $product->rights_information()->detach();
-        foreach ($rightsInformationArray as $rightsInformationEntity) {
-            $product->rights_information()->attach($rightsInformationEntity->id);
-            $rightsInformationEntity->product_id = $product->id;
-            $rightsInformationEntity->save();
+        if (isset($rightsInformationArray)) {
+            $product->rights_information()->detach();
+
+            foreach ($rightsInformationArray as $rightsInformationEntity) {
+                $product->rights_information()->attach($rightsInformationEntity->id);
+                $rightsInformationEntity->product_id = $product->id;
+                $rightsInformationEntity->save();
+            }
         }
 
         $productionInfo->product_id = $product->id;
